requirements.txt
requests==2.28.1
PIllow==9.2.0
-f https://download.pytorch.org/whl/cpu
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
